fix(core): Report no-op executor tasks as cancelled (JAVA-628)
NoOpSentryExecutorService returned a FutureTask that is never run and
never cancelled, so a caller had no way to tell a dropped task from a
queued one: isCancelled() and isDone() both stay false forever, and get()
blocks until it times out.

Return an already-cancelled Future instead, matching what
SentryExecutorService already does when its work queue is full, and state
the requirement on ISentryExecutorService so future implementations honour
it. CancelledFuture moves out of SentryExecutorService so both services
share one implementation.

An ActivityLifecycleIntegration test asserted a freshly scheduled ttfd
timeout was not cancelled while running on the no-op executor, so it was
really pinning the old pending-forever behaviour rather than the
scheduling it names. It now uses DeferredExecutorService, which queues.
